added autoscroll to message container for new messages received
